refactor [#237] 이벤트 관련 enum 객체 분리
