+++ CONTACTO-iOS/CONTACTO-iOS/Global/Resources/Amplitude/Event/EventName.swift
@@ -1,0 +1,139 @@
+//
+//  EventName.swift
+//  CONTACTO-iOS
+//
+//  Created by 장아령 on 4/10/25.
+//
+
+import Foundation
+
+
+public enum EventName: String {
+    // Onboading
+    case VIEW_ONBOARDING1
+    case CLICK_ONBOARDING1_NEXT
+    case VIEW_ONBOARDING2
+    case CLICK_ONBOARDING2_NEXT
+    case VIEW_ONBOARDING3
+    case CLICK_ONBOARDING3_NEXT
+    case VIEW_ONBOARDING4
+    case CLICK_ONBOARDING4_NEXT
+    case VIEW_ONBOARDING5
+    case CLICK_ONBOARDING5_NEXT
+    case VIEW_ONBOARDING6
+    case CLICK_ONBOARDING6_NEXT
+        
+    // Login
+    case VIEW_LOGIN
+    case VIEW_NOACCOUNT
+    case VIEW_INCORRECT
+    case VIEW_SIGNUP
+    case VIEW_EMAIL_CODE
+    case VIEW_SET_PASSWORD
+    case VIEW_SEND_CODE
+    case VIEW_RESET_PASSWORD
+    case VIEW_INPUT_NAME
+    case VIEW_HOME
+    case VIEW_HOME_TUTORIAL
+    case VIEW_MATCH
+
+    case CLICK_LOGIN_CONTINUE
+    case CLICK_LOGIN_CREATE
+    case CLICK_LOGIN_NEEDHELP
+    case CLICK_LOGIN_BUTTON
+    case CLICK_LOGIN_FIRST_STEP
+    
+    case CLICK_NOACCOUNT_CONTINUE
+    case CLICK_NOACCOUNT_FORGET
+    
+    case CLICK_INCORRECT_LOGIN
+    case CLICK_INCORRECT_FORGET
+    
+    case CLICK_SIGNUP_AGREE
+    case CLICK_SIGNUP_AGREE_DETAIL
+    case CLICK_SIGNUP_CONTINUE
+    case CLICK_SIGNUP_BACK
+    
+    case CLICK_EMAIL_CODE_NEXT
+    case CLICK_EMAIL_CODE_RESEND
+    
+    case CLICK_SET_PASSWORD_NEXT
+    
+    case CLICK_SEND_CODE_CONTINUE
+    case CLICK_SEND_CODE_FORGET
+    
+    case CLICK_RESET_PASSWORD_NEXT
+    case CLICK_INPUT_NAME_CONTINUE
+    case CLICK_INPUT_NAME_FORGET
+    case CLICK_INPUT_NAME_GO_TO_LOGIN
+    
+    case CLICK_HOME_BACK
+    case CLICK_HOME_NEXT
+    case CLICK_HOME_PROFILE
+    case VIEW_HOME_EMPTY
+    case CLICK_HOME_YES
+    case CLICK_HOME_NO
+    case CLICK_HOME_REVERT
+    
+    case SUCCESS_LOGIN
+    case TEST
+    
+    // Detail
+    case VIEW_DETAIL
+    case SCROLL_DETAIL
+    case CLICK_DETAIL_BACK
+    case CLICK_DETAIL_INSTA
+    case CLICK_DETAIL_WEB
+    case CLICK_DETAIL_BLOCK_YES
+    case CLICK_DETAIL_BLOCK_NO
+    case CLICK_DETAIL_REPORT_YES
+    case CLICK_DETAIL_REPORT_NO
+    
+    // Edit
+    case VIEW_EDIT
+    case SCROLL_EDIT
+    case CLICK_EDIT_PROFILE_EDIT
+    case CLICK_EDIT_PREVIEW
+    case CLICK_EDIT_NAME
+    case CLICK_EDIT_PORTFOLIO
+    case CLICK_EDIT_PORTFOLIO_DELETE
+    case CLICK_EDIT_TALENT
+    case CLICK_EDIT_DESCRIPTION
+    case CLICK_EDIT_PURPOSE
+    case CLICK_EDIT_INSTA
+    case CLICK_EDIT_WEB
+    case CLICK_EDIT_SAVE
+    case CLICK_EDIT_EDITSTART
+    case VIEW_PREVIEW
+    
+    
+    // Home
+    case CLICK_MATCH_SEND
+
+    // Chat
+    case VIEW_CHAT
+    case VIEW_EMPTY
+    case SCROLL_CHAT
+    case SCROLL_CHATROOM
+    case CLICK_CHAT
+    case VIEW_CHATROOM
+    case CLICK_CHATROOM_PLUS
+    case CLICK_CHATROOM_SEND
+    case CLICK_CHATROOM_TRANS_ON
+    case CLICK_CHATROOM_TRANS_OFF
+    case CLICK_CHATROOM_TRANSLATE_CHOOSE
+    case CLICK_CHATROOM_TRANSLATE_LANGUAGE
+    case CLICK_CHATROOM_BACK
+
+    // Info
+    case VIEW_INFO
+    case CLICK_INFO_HELP
+    case CLICK_INFO_COMMUNITY
+    case CLICK_INFO_PRIVACY
+    case CLICK_INFO_LOGOUT
+    case CLICK_INFO_LOGOUT_YES
+    case CLICK_INFO_LOGOUT_NO
+    case CLICK_INFO_DELETE
+    case CLICK_INFO_DELETE_YES
+    case CLICK_INFO_DELETE_NO
+}
