+++ CONTACTO-iOS/CONTACTO-iOS/Global/Resources/Amplitude/Event/EventView.swift
@@ -1,0 +1,19 @@
+//
+//  EventView.swift
+//  CONTACTO-iOS
+//
+//  Created by 장아령 on 4/10/25.
+//
+
+import Foundation
+
+public enum EventView: String {
+    case ONBOARDING = "회원가입"
+    case LOGIN = "로그인"
+    case HOME = "홈"
+    case DETAIL = "상세조회"
+    case CHAT = "채팅"
+    case EDIT = "편집"
+    case INFO = "인포"
+    case PUSH = "푸시 알림"
+}
